refactor: common => core
